EMNavSetup
API connection to NDI Aurora EM sensing. This script will download the required libraries for the API. It will then go through the process of setting up the connection between the python API and the EM sensing. Next it will define the setting for serial commutation. From here we will handle the ports and initialize them to the proper tool. We are now able to begin tracking and we have included a snipping of code that shows live data accusation from the sensor.