two-way python/arduino serial communication
Arduino stores serial instructions on a queue. The issues about timing between python and Arduino was due to python. Sending commands must be a priory in python and everything else is less important for communication to work properly. With rough timing sketch I was able to time the latency between sending a command and Arduino starting to execute the command and it is usually between 1-2 ms.